+++ docs/search/all_2.js
@@ -3,9 +3,8 @@ var searchData=
   ['cameraposition_6',['cameraPosition',['../classsmall3d_1_1_renderer.html#a98e23bfad9181afafcf6f64655698c35',1,'small3d::Renderer']]],
   ['camerarotation_7',['cameraRotation',['../classsmall3d_1_1_renderer.html#ab560a8768b39a18d938f2a2de943c618',1,'small3d::Renderer']]],
   ['clearbuffers_8',['clearBuffers',['../classsmall3d_1_1_renderer.html#acd6b563ad9f26558659cb335b06fc95d',1,'small3d::Renderer::clearBuffers(Model &amp;model) const'],['../classsmall3d_1_1_renderer.html#a2b636d995107d5dce4c20c465d0f6187',1,'small3d::Renderer::clearBuffers(SceneObject &amp;sceneObject) const']]],
-  ['clearscreen_9',['clearScreen',['../classsmall3d_1_1_renderer.html#af85d1146e340e49cd1b568f7f3cce8cd',1,'small3d::Renderer::clearScreen() const'],['../classsmall3d_1_1_renderer.html#ad89aca86a938646fcf2273aa70dce33a',1,'small3d::Renderer::clearScreen(const glm::vec4 &amp;colour)']]],
-  ['contains_10',['contains',['../classsmall3d_1_1_bounding_box_set.html#a4b5b507e6dae95b55943ec09d1d2d3bb',1,'small3d::BoundingBoxSet::contains()'],['../classsmall3d_1_1_scene_object.html#ae545d15364b5ecb0f4bc0d5cb5629ba4',1,'small3d::SceneObject::contains()']]],
-  ['containscorners_11',['containsCorners',['../classsmall3d_1_1_bounding_box_set.html#a3f168bc8505d561601e94d9b77c4b07c',1,'small3d::BoundingBoxSet::containsCorners()'],['../classsmall3d_1_1_scene_object.html#a0cdbb5e24f06458dc5451acec3e373b7',1,'small3d::SceneObject::containsCorners()']]],
-  ['createrectangle_12',['createRectangle',['../classsmall3d_1_1_renderer.html#acaa5ae15a3a21aaf43bbce93b5d3cdd7',1,'small3d::Renderer']]],
-  ['changelog_13',['changelog',['../md_changelog.html',1,'(Global Namespace)'],['../changelog.html',1,'(Global Namespace)']]]
+  ['contains_9',['contains',['../classsmall3d_1_1_bounding_box_set.html#a4b5b507e6dae95b55943ec09d1d2d3bb',1,'small3d::BoundingBoxSet::contains()'],['../classsmall3d_1_1_scene_object.html#ae545d15364b5ecb0f4bc0d5cb5629ba4',1,'small3d::SceneObject::contains()']]],
+  ['containscorners_10',['containsCorners',['../classsmall3d_1_1_bounding_box_set.html#a3f168bc8505d561601e94d9b77c4b07c',1,'small3d::BoundingBoxSet::containsCorners()'],['../classsmall3d_1_1_scene_object.html#a0cdbb5e24f06458dc5451acec3e373b7',1,'small3d::SceneObject::containsCorners()']]],
+  ['createrectangle_11',['createRectangle',['../classsmall3d_1_1_renderer.html#acaa5ae15a3a21aaf43bbce93b5d3cdd7',1,'small3d::Renderer']]],
+  ['changelog_12',['changelog',['../md_changelog.html',1,'(Global Namespace)'],['../changelog.html',1,'(Global Namespace)']]]
 ];