+++ docs/search/all_e.js
@@ -1,9 +1,10 @@
 var searchData=
 [
-  ['small3d_76',['small3d',['../index.html',1,'']]],
-  ['scale_77',['scale',['../classsmall3d_1_1_model.html#a39352fd779afaef6ee5ef398a0481b31',1,'small3d::Model']]],
-  ['sceneobject_78',['SceneObject',['../classsmall3d_1_1_scene_object.html',1,'small3d::SceneObject'],['../classsmall3d_1_1_scene_object.html#a1bd665f32145f9b16fc89df98c8713a9',1,'small3d::SceneObject::SceneObject(const std::string name, const std::string modelPath, const std::string &amp;modelMeshName, const uint32_t boundingBoxSubdivisions=0)'],['../classsmall3d_1_1_scene_object.html#a2e9dfda6e679bbacce88ffdb63685e7e',1,'small3d::SceneObject::SceneObject(const std::string name, const std::string modelPath, const int numFrames=1, const int startFrameIndex=1, const uint32_t boundingBoxSubdivisions=0)']]],
-  ['sceneobject_2ehpp_79',['SceneObject.hpp',['../_scene_object_8hpp.html',1,'']]],
+  ['small3d_75',['small3d',['../index.html',1,'']]],
+  ['scale_76',['scale',['../classsmall3d_1_1_model.html#a39352fd779afaef6ee5ef398a0481b31',1,'small3d::Model']]],
+  ['sceneobject_77',['SceneObject',['../classsmall3d_1_1_scene_object.html',1,'small3d::SceneObject'],['../classsmall3d_1_1_scene_object.html#a1bd665f32145f9b16fc89df98c8713a9',1,'small3d::SceneObject::SceneObject(const std::string name, const std::string modelPath, const std::string &amp;modelMeshName, const uint32_t boundingBoxSubdivisions=0)'],['../classsmall3d_1_1_scene_object.html#a2e9dfda6e679bbacce88ffdb63685e7e',1,'small3d::SceneObject::SceneObject(const std::string name, const std::string modelPath, const int numFrames=1, const int startFrameIndex=1, const uint32_t boundingBoxSubdivisions=0)']]],
+  ['sceneobject_2ehpp_78',['SceneObject.hpp',['../_scene_object_8hpp.html',1,'']]],
+  ['setbackgroundcolour_79',['setBackgroundColour',['../classsmall3d_1_1_renderer.html#ab68576dfc33356d43722e83fadec55f4',1,'small3d::Renderer']]],
   ['setframedelay_80',['setFrameDelay',['../classsmall3d_1_1_scene_object.html#ac5597d21a46f7208fc92377f9cc8b631',1,'small3d::SceneObject']]],
   ['small3d_81',['small3d',['../namespacesmall3d.html',1,'']]],
   ['sound_82',['Sound',['../classsmall3d_1_1_sound.html',1,'small3d::Sound'],['../classsmall3d_1_1_sound.html#a3bf41af25f517dc9489d795402b18219',1,'small3d::Sound::Sound()'],['../classsmall3d_1_1_sound.html#a5c0dcc69eba4f4f1993d3d8aa50b28f7',1,'small3d::Sound::Sound(const std::string soundFilePath)'],['../classsmall3d_1_1_sound.html#abe9fffa807136a82dba901f6c1b9aeb7',1,'small3d::Sound::Sound(const Sound &amp;other) noexcept'],['../classsmall3d_1_1_sound.html#aa6e96196c84086a30c9d02aa1a4aa358',1,'small3d::Sound::Sound(const Sound &amp;&amp;other) noexcept']]],